test(integration): 🔧 restructure integration test organization
Refactor the integration test structure by renaming and reorganizing namespaces and base classes. This enhances consistency across test classes and aligns with a new naming convention, improving the overall clarity and maintainability of the test suite. Additionally, the organization of test classes has been refined to better reflect their functionality.
