+++ src/Tests/Integration/Connections/DirectConnectionTests.cs
@@ -2,91 +2,89 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Void.Minecraft.Network;
+using Void.Tests.Integration.Base;
 using Void.Tests.Integration.Sides.Clients;
-using Void.Tests.Integration.Sides.Proxies;
 using Void.Tests.Integration.Sides.Servers;
 using Xunit;
 
-namespace Void.Tests.Integration.Connections.Units;
+namespace Void.Tests.Integration.Connections;
 
-public class ProxiedConnectionTests(ProxiedConnectionTests.Fixture fixture) : ConnectionUnitBase, IClassFixture<ProxiedConnectionTests.Fixture>
+public class DirectConnectionTests(DirectConnectionTests.Fixture fixture) : IntegrationUnitBase, IClassFixture<DirectConnectionTests.Fixture>
 {
-    private const int ProxyPort = 35000;
-    private const int ServerPort = 35001;
-    private const string ExpectedText = "hello proxied void!";
+    private const int ServerPort = 25000;
+    private const string ExpectedText = "hello void!";
 
-    [ProxiedFact]
-    public async Task MccConnectsToPaperServerThroughProxy()
+    [DirectFact]
+    public async Task MccConnectsToPaperServer()
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ProxyPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
+            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ServerPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MinecraftConsoleClient, fixture.VoidProxy, fixture.PaperServer);
+        }, fixture.MinecraftConsoleClient, fixture.PaperServer);
     }
 
-    [ProxiedTheory]
+    [DirectTheory]
     [MemberData(nameof(MinecraftConsoleClient.SupportedVersions), MemberType = typeof(MinecraftConsoleClient))]
-    public async Task MccConnectsToPaperServerThroughProxy_WithProtocolVersion(ProtocolVersion protocolVersion)
+    public async Task MccConnectsToPaperServer_WithProtocolVersion(ProtocolVersion protocolVersion)
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ProxyPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
+            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ServerPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MinecraftConsoleClient, fixture.VoidProxy, fixture.PaperServer);
+        }, fixture.MinecraftConsoleClient, fixture.PaperServer);
     }
 
-    [ProxiedFact]
-    public async Task MineflayerConnectsToPaperServerThroughProxy()
+    [DirectFact]
+    public async Task MineflayerConnectsToPaperServer()
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ProxyPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
+            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ServerPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MineflayerClient, fixture.VoidProxy, fixture.PaperServer);
+        }, fixture.MineflayerClient, fixture.PaperServer);
     }
 
-    [ProxiedTheory]
+    [DirectTheory]
     [MemberData(nameof(MineflayerClient.SupportedVersions), MemberType = typeof(MineflayerClient))]
-    public async Task MineflayerConnectsToPaperServerThroughProxy_WithProtocolVersion(ProtocolVersion protocolVersion)
+    public async Task MineflayerConnectsToPaperServer_WithProtocolVersion(ProtocolVersion protocolVersion)
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ProxyPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
+            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ServerPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MineflayerClient, fixture.VoidProxy, fixture.PaperServer);
+        }, fixture.MineflayerClient, fixture.PaperServer);
     }
 
-    public class Fixture : ConnectionFixtureBase, IAsyncLifetime
+    public class Fixture : IntegrationFixtureBase, IAsyncLifetime
     {
-        public Fixture() : base(nameof(ProxiedConnectionTests))
+        public Fixture() : base(nameof(DirectConnectionTests))
         {
         }
 
         public MinecraftConsoleClient MinecraftConsoleClient { get => field ?? throw new InvalidOperationException($"{nameof(MinecraftConsoleClient)} is not initialized."); set; }
         public MineflayerClient MineflayerClient { get => field ?? throw new InvalidOperationException($"{nameof(MineflayerClient)} is not initialized."); set; }
         public PaperServer PaperServer { get => field ?? throw new InvalidOperationException($"{nameof(PaperServer)} is not initialized."); set; }
-        public VoidProxy VoidProxy { get => field ?? throw new InvalidOperationException($"{nameof(VoidProxy)} is not initialized."); set; }
 
         public async Task InitializeAsync()
         {
@@ -95,7 +93,6 @@ public class ProxiedConnectionTests(ProxiedConnectionTests.Fixture fixture) : Co
             MinecraftConsoleClient = await MinecraftConsoleClient.CreateAsync(_workingDirectory, _httpClient, cancellationToken: cancellationTokenSource.Token);
             MineflayerClient = await MineflayerClient.CreateAsync(_workingDirectory, _httpClient, cancellationToken: cancellationTokenSource.Token);
             PaperServer = await PaperServer.CreateAsync(_workingDirectory, _httpClient, port: ServerPort, cancellationToken: cancellationTokenSource.Token);
-            VoidProxy = await VoidProxy.CreateAsync(_workingDirectory, targetServer: $"localhost:{ServerPort}", proxyPort: ProxyPort, cancellationToken: cancellationTokenSource.Token);
         }
 
         public async Task DisposeAsync()
@@ -108,9 +105,6 @@ public class ProxiedConnectionTests(ProxiedConnectionTests.Fixture fixture) : Co
 
             if (PaperServer is not null)
                 await PaperServer.DisposeAsync();
-
-            if (VoidProxy is not null)
-                await VoidProxy.DisposeAsync();
         }
     }
 }
