+++ src/Tests/Integration/Connections/ProxiedConnectionTests.cs
@@ -2,88 +2,92 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Void.Minecraft.Network;
+using Void.Tests.Integration.Base;
 using Void.Tests.Integration.Sides.Clients;
+using Void.Tests.Integration.Sides.Proxies;
 using Void.Tests.Integration.Sides.Servers;
 using Xunit;
 
-namespace Void.Tests.Integration.Connections.Units;
+namespace Void.Tests.Integration.Connections;
 
-public class DirectConnectionTests(DirectConnectionTests.Fixture fixture) : ConnectionUnitBase, IClassFixture<DirectConnectionTests.Fixture>
+public class ProxiedConnectionTests(ProxiedConnectionTests.Fixture fixture) : IntegrationUnitBase, IClassFixture<ProxiedConnectionTests.Fixture>
 {
-    private const int ServerPort = 25000;
-    private const string ExpectedText = "hello void!";
+    private const int ProxyPort = 35000;
+    private const int ServerPort = 35001;
+    private const string ExpectedText = "hello proxied void!";
 
-    [DirectFact]
-    public async Task MccConnectsToPaperServer()
+    [ProxiedFact]
+    public async Task MccConnectsToPaperServerThroughProxy()
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ServerPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
+            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ProxyPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MinecraftConsoleClient, fixture.PaperServer);
+        }, fixture.MinecraftConsoleClient, fixture.VoidProxy, fixture.PaperServer);
     }
 
-    [DirectTheory]
+    [ProxiedTheory]
     [MemberData(nameof(MinecraftConsoleClient.SupportedVersions), MemberType = typeof(MinecraftConsoleClient))]
-    public async Task MccConnectsToPaperServer_WithProtocolVersion(ProtocolVersion protocolVersion)
+    public async Task MccConnectsToPaperServerThroughProxy_WithProtocolVersion(ProtocolVersion protocolVersion)
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ServerPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
+            await fixture.MinecraftConsoleClient.SendTextMessageAsync($"localhost:{ProxyPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MinecraftConsoleClient, fixture.PaperServer);
+        }, fixture.MinecraftConsoleClient, fixture.VoidProxy, fixture.PaperServer);
     }
 
-    [DirectFact]
-    public async Task MineflayerConnectsToPaperServer()
+    [ProxiedFact]
+    public async Task MineflayerConnectsToPaperServerThroughProxy()
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ServerPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
+            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ProxyPort}", ProtocolVersion.MINECRAFT_1_20_3, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MineflayerClient, fixture.PaperServer);
+        }, fixture.MineflayerClient, fixture.VoidProxy, fixture.PaperServer);
     }
 
-    [DirectTheory]
+    [ProxiedTheory]
     [MemberData(nameof(MineflayerClient.SupportedVersions), MemberType = typeof(MineflayerClient))]
-    public async Task MineflayerConnectsToPaperServer_WithProtocolVersion(ProtocolVersion protocolVersion)
+    public async Task MineflayerConnectsToPaperServerThroughProxy_WithProtocolVersion(ProtocolVersion protocolVersion)
     {
         var expectedText = $"{ExpectedText} test #{Random.Shared.Next()}";
         using var cancellationTokenSource = new CancellationTokenSource(Timeout);
 
         await LoggedExecutorAsync(async () =>
         {
-            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ServerPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
+            await fixture.MineflayerClient.SendTextMessageAsync($"localhost:{ProxyPort}", protocolVersion, expectedText, cancellationTokenSource.Token);
             await fixture.PaperServer.ExpectTextAsync(expectedText, lookupHistory: true, cancellationTokenSource.Token);
 
             Assert.Contains(fixture.PaperServer.Logs, line => line.Contains(expectedText));
-        }, fixture.MineflayerClient, fixture.PaperServer);
+        }, fixture.MineflayerClient, fixture.VoidProxy, fixture.PaperServer);
     }
 
-    public class Fixture : ConnectionFixtureBase, IAsyncLifetime
+    public class Fixture : IntegrationFixtureBase, IAsyncLifetime
     {
-        public Fixture() : base(nameof(DirectConnectionTests))
+        public Fixture() : base(nameof(ProxiedConnectionTests))
         {
         }
 
         public MinecraftConsoleClient MinecraftConsoleClient { get => field ?? throw new InvalidOperationException($"{nameof(MinecraftConsoleClient)} is not initialized."); set; }
         public MineflayerClient MineflayerClient { get => field ?? throw new InvalidOperationException($"{nameof(MineflayerClient)} is not initialized."); set; }
         public PaperServer PaperServer { get => field ?? throw new InvalidOperationException($"{nameof(PaperServer)} is not initialized."); set; }
+        public VoidProxy VoidProxy { get => field ?? throw new InvalidOperationException($"{nameof(VoidProxy)} is not initialized."); set; }
 
         public async Task InitializeAsync()
         {
@@ -92,6 +96,7 @@ public class DirectConnectionTests(DirectConnectionTests.Fixture fixture) : Conn
             MinecraftConsoleClient = await MinecraftConsoleClient.CreateAsync(_workingDirectory, _httpClient, cancellationToken: cancellationTokenSource.Token);
             MineflayerClient = await MineflayerClient.CreateAsync(_workingDirectory, _httpClient, cancellationToken: cancellationTokenSource.Token);
             PaperServer = await PaperServer.CreateAsync(_workingDirectory, _httpClient, port: ServerPort, cancellationToken: cancellationTokenSource.Token);
+            VoidProxy = await VoidProxy.CreateAsync(_workingDirectory, targetServer: $"localhost:{ServerPort}", proxyPort: ProxyPort, cancellationToken: cancellationTokenSource.Token);
         }
 
         public async Task DisposeAsync()
@@ -104,6 +109,9 @@ public class DirectConnectionTests(DirectConnectionTests.Fixture fixture) : Conn
 
             if (PaperServer is not null)
                 await PaperServer.DisposeAsync();
+
+            if (VoidProxy is not null)
+                await VoidProxy.DisposeAsync();
         }
     }
 }
